chore: add nav menu group
