+++ src/theme-default/components/Nav/index.module.scss
@@ -201,3 +201,6 @@
     padding-left: 32px;
   }
 }
+
+.menuGroup {
+}
